requirements.txt
openai==0.27.0
langchain==0.0.320
chromadb==0.4.1
tqdm==4.66.1
python-dotenv==1.0.0
tiktoken==0.4.0
python-docx>=0.8.11
